chore: Update npm dependencies
This commit updates the npm dependencies in the project. Specifically, it adds the "react-router-dom" package with version 6.24.1. This package is required for implementing routing functionality in the application. Additionally, it updates the "react-router" package to version 6.24.1 and adds the "@remix-run/router" package with version 1.17.1 as a dependency for "react-router-dom". These updates ensure compatibility and stability with the latest versions of React and other related packages.

Note: This commit message follows the established convention of starting with a verb in the imperative mood, followed by a brief description of the changes.